Allow interacting with the casting table tank using fluid items
If the fluid item can interact (either filling or emptying the table tank), it will
If unable to interact, it will attempt to be placed on the table
This is based on the assumption you won't have a fluid item that contains fluid which you wish to cast upon (e.g. filled bucket or filled tool). If you really need to for some reason, use a scorched table or a hopper